rename component file
